+++ sdk/resourcemanager/computeschedule/armcomputeschedule/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 1.0.0 (2025-01-24)
+### Breaking Changes
+
+- Type of `OperationErrorDetails.ErrorDetails` has been changed from `*time.Time` to `*string`
+
+### Features Added
+
+- New field `AzureOperationName`, `Timestamp` in struct `OperationErrorDetails`
+- New field `Timezone` in struct `ResourceOperationDetails`
+- New field `Deadline`, `Timezone` in struct `Schedule`
+
+
 ## 0.1.0 (2024-09-27)
 ### Other Changes